// src/generated/client/models/IntentType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Metamind API
 * Brain spec for Metamind.
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum IntentType {
    NORMAL = 'NORMAL',
    CONFUSED = 'CONFUSED',
    REDIRECT = 'REDIRECT',
    DEFAULT = 'DEFAULT'
}

export function IntentTypeFromJSON(json: any): IntentType {
    return IntentTypeFromJSONTyped(json, false);
}

export function IntentTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): IntentType {
    return json as IntentType;
}

export function IntentTypeToJSON(value?: IntentType | null): any {
    return value as any;
}